geotree in working order
